+++ extensions/sevalla/CHANGELOG.md
@@ -1,3 +1,12 @@
 # Sevalla Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Search Databases
+    1. View Database Details
+    2. Delete Database
+    3. Create Database
+- Search Static Sites
+    1. View Static Site Details
+        - View Deployment Details
+    2. Delete Static Site